+++ app/src/androidTest/java/com/averycorp/prismtask/CloudIdOrphanHealerScenarioTest.kt
@@ -91,7 +91,6 @@ class CloudIdOrphanHealerScenarioTest {
             medicationSlotDao = database.medicationSlotDao(),
             medicationSlotOverrideDao = database.medicationSlotOverrideDao(),
             medicationTierStateDao = database.medicationTierStateDao(),
-            medicationMarkDao = database.medicationMarkDao(),
             logger = logger
         )
     }