+++ app/src/androidTest/java/com/averycorp/prismtask/CloudIdOrphanHealerTest.kt
@@ -85,7 +85,6 @@ class CloudIdOrphanHealerTest {
             medicationSlotDao = database.medicationSlotDao(),
             medicationSlotOverrideDao = database.medicationSlotOverrideDao(),
             medicationTierStateDao = database.medicationTierStateDao(),
-            medicationMarkDao = database.medicationMarkDao(),
             logger = logger
         )
     }